+++ CHANGELOG.md
@@ -29,4 +29,4 @@
 
 * Added order lines for Klarna Pay Later payment method
 * Added order lines for AfterPay payment method
-* Fixed incorrect amount value
+* Fixed incorrect amount value of order